+++ bootique-di/src/main/java/io/bootique/di/spi/FieldInjectingProvider.java
@@ -8,6 +8,7 @@ import javax.inject.Inject;
 import javax.inject.Provider;
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Field;
+import java.lang.reflect.Modifier;
 import java.lang.reflect.Type;
 
 class FieldInjectingProvider<T> extends MemberInjectingProvider<T> {
@@ -20,13 +21,17 @@ class FieldInjectingProvider<T> extends MemberInjectingProvider<T> {
     protected void injectMembers(T object, Class<?> type) {
 
         // bail on recursion stop condition
-        if (type == null) {
+        if (type == Object.class) {
             return;
         }
 
         injectMembers(object, type.getSuperclass());
 
         for (Field field : type.getDeclaredFields()) {
+            if(Modifier.isStatic(field.getModifiers())) {
+                // skip static fields completely
+                continue;
+            }
             Inject inject = field.getAnnotation(Inject.class);
             if (inject != null) {
                 injectMember(object, field, getQualifier(field));